jstation: move midi related sources to a midi module
